Postcss 8 (#35)
* update to postcss 8

This is just a straightforward version bump, using the `Root:` style
postcss 8 plugin. I tried to switch to using `Declaration:`, but we get
in an infinite replacement loop. I could remedy it with some extra
checks in the `transition` transform, but I'm not confident that that
were all the possible infinite loop cases, so I went back to this
single-pass style.

* up rollup

* ci: bump matrix

* update engines

* remove Root func and use Declaration

Co-authored-by: Renée Kooi <renee@kooi.me>